% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignWinner.R
\name{assignWinner}
\alias{assignWinner}
\title{assignWinner}
\usage{
assignWinner(RoundData, Subs, teams, type = "normal")
}
\arguments{
\item{RoundData}{a \code{data.frame} that includes the team matchups}

\item{Subs}{a \code{data.frame} that has t1ID, t2ID and Pred}

\item{teams}{a \code{data.frame} that has the ids and team names for each team in the tournament}
}
