% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBracket.R
\name{createBracket}
\alias{createBracket}
\title{createBracket}
\usage{
createBracket(seeds, Subs, teams, season = 2019, type = "normal")
}
\arguments{
\item{seeds}{a \code{data.frame} of the seeds in the tournament}

\item{Subs}{a \code{data.frame} that has t1ID, t2ID and Pred}

\item{teams}{a \code{data.frame} that has the ids and team names for each team in the tournament}

\item{season}{specify the bracket season}

\item{type}{specify the type of bracket you would like to build. \code{type = 'normal'} will return always select the team with the highest probability as the winner. \code{type = 'simulate'} will randomly assign the winner based on the probabilities}
}
