% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_2002_19}
\alias{Data_2002_19}
\title{Tournament Data 2002 - 2019}
\format{
A data frame with the following columns:

\describe{

\item{team}{test}
\item{year}{test}
\item{region}{test}
\item{seed}{test2}
\item{play.in}{...}
\item{rd_1}{...}
\item{rd_2}{...}
\item{rd_3}{...}
\item{rd_4}{...}
\item{rd_5}{...}
\item{rd_6}{...}
\item{TPPG}{...}
\item{TPPG_last3}{...}
\item{FGPG}{...}
\item{FGPG_last3}{...}
\item{FTAPG}{...}
\item{FTAPG_last3}{...}
\item{FTMPP}{...}
\item{FTMPP_last3}{...}
\item{ORP}{...}
\item{ORP_last3}{...}
\item{ORPG}{...}
\item{ORPG_last3}{...}
\item{DRP}{...}
\item{DRP_last3}{...}
\item{DRPG}{...}
\item{DRPG_last3}{...}
\item{APG}{...}
\item{APG_last3}{...}
\item{FPG}{...}
\item{FPG_last3}{...}
\item{ASM}{...}
\item{ASM_last3}{...}
\item{ATR}{...}
\item{ATR_last3}{...}
\item{OE}{...}
\item{OE_last3}{...}
\item{DE}{...}
\item{DE_last3}{...}

}
}
\source{
\doi{test}
}
\usage{
Data_2002_19
}
\description{
Description
}
\keyword{datasets}
