% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(DF)
}
\arguments{
\item{DF}{detailed results data (either from the tournament or regular season)}
}
\description{
Creates a more useful dataset for modeling purposes. Takes in detailed results data
}
