% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSubmission.R
\name{processSubmission}
\alias{processSubmission}
\title{processSubmission}
\usage{
processSubmission(subm, predictorDF, teams, year = 2018)
}
\arguments{
\item{subm}{the raw file for the kaggle submission}

\item{teams}{a \code{data.frame} that has the ids and team names for each team in the tournament}

\item{year}{specify the year you would like to process}
}
\description{
This function will set up a \code{data.frame} that can be passed to your model for predictions. Note: this function assumes that your team 2 variables are the same as team 1 and are named T2{variablename}
}
