% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{The Shiny App for Interactive table}
\usage{
run_app()
}
\value{
It will jump to a new web window to show the basic shiny app.
\dontrun{
run_app()
}
}
\description{
This is a basic shiny app to run an interactive table to explore the data.
}
