% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_better_app}
\alias{run_better_app}
\title{The Better App}
\usage{
run_better_app()
}
\value{
It will jump to a new web window to show the better shiny app.
\dontrun{
run_better_app()
}
}
\description{
This is a better version of my shiny app, including two interactive plots and one interactive table.
}
